Remove HTTPUtils
