+++ lib/timber/config.ex
@@ -22,5 +22,5 @@ defmodule Timber.Config do
   end
 
   def transport,
-    do: Application.get_env(@env_key, :transport, Timber.Transports.IODevice)
+    do: Application.get_env(@env_key, :transport)
 end
